+++ xdbit_funcs/images/__init__.py
@@ -3,6 +3,6 @@ from .image_processing import *
 from .segment import *
 from .features import *
 from .registration import *
-from .container import *
+#from .container import *
 from .interactive import *
 from .zarr import *
